Move isGuardIf to where it belongs (according to the comment on top of JavaAstUtils)
